+++ app.py
@@ -1,7 +1,7 @@
 import streamlit as st
 from openai import OpenAI
-import PyPDF2
 import re
+import core
 
 # -----------------------------
 # App Config
@@ -61,234 +61,34 @@ st.markdown("""
 """, unsafe_allow_html=True)
 
 # -----------------------------
-# Helper: PDF to text
+# LLM Interaction Helpers (using core)
 # -----------------------------
-def extract_text_from_pdf(uploaded_pdf):
-    reader = PyPDF2.PdfReader(uploaded_pdf)
-    text = ""
-    for page in reader.pages:
-        page_text = page.extract_text()
-        if page_text:
-            text += page_text + "\n"
-    if not text.strip():
-        raise ValueError("No extractable text found. The PDF may be image-only or empty.")
-    return text
 
-# -----------------------------
-# Compress text for token safety
-# -----------------------------
-def compress_text(text, limit=6000):
+def get_remedies_advice(judgment_text, language):
     """
-    Truncates text by taking the first and last portions, 
-    ensuring we break at sentence or paragraph boundaries.
+    Call LLM to get remedies for this judgment
     """
-    if len(text) <= limit:
-        return text
-
-    half = limit // 2
+    prompt = core.build_remedies_prompt(core.compress_text(judgment_text), language)
     
-    # 1. Process Head: Try to find the last sentence/para boundary in the first half
-    head_raw = text[:half]
-    # Look for . ! ? followed by space/newline, or multiple newlines
-    match_head = list(re.finditer(r'([.!?]\s+|\n+)', head_raw))
-    if match_head:
-        head_end = match_head[-1].end()
-        # Only truncate at the boundary if it doesn't discard too much (at least 70% of half)
-        if head_end > half * 0.7:
-            head = head_raw[:head_end]
-        else:
-            head = head_raw
-    else:
-        # Fallback: search for last space to avoid cutting a word
-        last_space = head_raw.rfind(' ')
-        head = head_raw[:last_space] if last_space != -1 else head_raw
-
-    # 2. Process Tail: Try to find the first sentence/para boundary in the last half
-    tail_raw = text[-half:]
-    match_tail = list(re.finditer(r'([.!?]\s+|\n+)', tail_raw))
-    if match_tail:
-        # Start after the first boundary found in the first 30% of the tail segment
-        tail_start = match_tail[0].end()
-        if tail_start < half * 0.3:
-            tail = tail_raw[tail_start:]
-        else:
-            tail = tail_raw
-    else:
-        # Fallback: search for first space
-        first_space = tail_raw.find(' ')
-        tail = tail_raw[first_space+1:] if first_space != -1 else tail_raw
-
-    return head.strip() + "\n\n... [TRUNCATED] ...\n\n" + tail.strip()
-
-# -----------------------------
-# Detect English leakage
-# -----------------------------
-def english_leakage_detected(output_text, threshold=5):
-    common = [" the ", " and ", " of ", " to ", " in ", " is ", " that ", " it ", " for ", " on "]
-    text_lower = " " + output_text.lower() + " "
-    count = sum(1 for w in common if w in text_lower)
-    return count >= threshold
-
-# -----------------------------
-# Build prompts
-# -----------------------------
-def build_prompt(safe_text, language):
-    return f"""
-You are LegalEase AI — an expert judicial-simplification and translation engine.
-
-MISSION:
-Convert the judgment text into a simple, citizen-friendly summary.
-
-INSTRUCTIONS:
-1. Extract ONLY the final judgment outcome.
-2. Remove all legal jargon and case history.
-3. Produce EXACTLY 3 bullet points.
-4. Write ONLY in {language}. ZERO English allowed if language ≠ English.
-5. Each bullet must be 1–2 very short sentences.
-6. No extra headings. No disclaimers.
-
-TEXT TO ANALYZE:
-{safe_text}
-
-OUTPUT REQUIRED:
-- 3 bullet points in {language} only
-"""
-
-def build_retry_prompt(safe_text, language):
-    return f"""
-Your previous answer included English. Now STRICTLY produce the answer ONLY in {language}.
-
-REQUIREMENTS:
-- Exactly 3 bullet points
-- VERY simple {language}
-- No English at all
-- No introductions, headings, or explanations
-
-TEXT:
-{safe_text}
-
-OUTPUT NOW:
-3 bullet points in {language} only.
-"""
-
-# -----------------------------
-# Remedies Advisor Functions
-# -----------------------------
-
-def build_remedies_prompt(judgment_text, language):
-    """
-    Ask LLM to analyze what remedies are available
-    based on the actual judgment content
-    """
-    return f"""
-You are a Legal Rights Advisor. Read this judgment and answer in SIMPLE format.
-
-JUDGMENT:
-{judgment_text}
-
-Answer ONLY these questions in {language}. Be practical and direct.
-
-1. What happened? (Who won and who lost; 1 sentence)
-2. Can the loser appeal? (Yes/No + reason; 1-2 sentences)
-3. Appeal timeline: How many days? (Just number)
-4. Appeal court: Which court should they go to? (Court name only)
-5. Cost estimate: Rough cost in rupees? (e.g., 5000-15000)
-6. First action: What should they do first? (1 sentence)
-7. Important deadline: What key deadline should they remember? (1 sentence)
-
-Output in numbered form like:
-1. ...\n2. ...\n3. ... etc.
-"""
-
-
-def parse_remedies_response(response_text):
-    """
-    Extract structured info from LLM response using flexible numbered-line parsing.
-    Supports multiple separators: . ) : - 
-    Handles both 5-section (old) and 7-section (new) formats.
-    """
-    import re
+    response = client.chat.completions.create(
+        model="meta-llama/llama-3.1-8b-instruct",
+        messages=[
+            {
+                "role": "system",
+                "content": "You are a helpful legal advisor. Answer questions about legal remedies in India."
+            },
+            {
+                "role": "user",
+                "content": prompt
+            }
+        ],
+        max_tokens=500,  # Longer for detailed answers
+        temperature=0.1,  # Low temp = more consistent
+    )
     
-    remedies = {
-        "what_happened": "",
-        "can_appeal": "",
-        "appeal_days": "",
-        "appeal_court": "",
-        "cost_estimate": "",
-        "cost": "",
-        "first_action": "",
-        "deadline": "",
-        "appeal_details": ""
-    }
-
-    text = response_text.strip()
-    if not text:
-        return remedies
-
-    # Detect all numbered sections (flexible separators: . ) : -)
-    # Only match 1-2 digit numbers to avoid matching content like "5000-10000"
-    pattern = r'^([1-9]\d?)\s*[.):‐-]\s*(.*?)$'
-    sections = {}
+    response_text = response.choices[0].message.content.strip()
+    remedies = core.parse_remedies_response(response_text)
     
-    for line in text.split('\n'):
-        match = re.match(pattern, line.strip())
-        if match:
-            num = int(match.group(1))
-            header = match.group(2).strip()
-            sections[num] = {"header": header, "content": ""}
-    
-    # Extract content for each section
-    lines = text.split('\n')
-    current_section = None
-    
-    for line in lines:
-        match = re.match(pattern, line.strip())
-        if match:
-            current_section = int(match.group(1))
-        elif current_section is not None and current_section in sections:
-            if line.strip():  # Only add non-empty lines
-                sections[current_section]["content"] += line.strip() + " "
-    
-    # Clean up content
-    for num in sections:
-        sections[num]["content"] = sections[num]["content"].strip()
-    
-    # Map sections to keys based on count
-    is_7section = len(sections) >= 7
-    
-    if is_7section:
-        if 1 in sections:
-            remedies["what_happened"] = sections[1]["content"]
-        if 2 in sections:
-            can_appeal_text = sections[2]["content"].lower()
-            remedies["can_appeal"] = "yes" if "yes" in can_appeal_text else "no"
-        if 3 in sections:
-            # Extract just the number from "30 days"
-            appeal_days_text = sections[3]["content"]
-            match = re.search(r'\d+', appeal_days_text)
-            remedies["appeal_days"] = match.group() if match else appeal_days_text
-        if 4 in sections:
-            remedies["appeal_court"] = sections[4]["content"]
-        if 5 in sections:
-            remedies["cost_estimate"] = sections[5]["content"]
-            remedies["cost"] = sections[5]["content"]  # Support both keys
-        if 6 in sections:
-            remedies["first_action"] = sections[6]["content"]
-        if 7 in sections:
-            remedies["deadline"] = sections[7]["content"]
-    else:
-        # 5-section format (old)
-        if 1 in sections:
-            remedies["what_happened"] = sections[1]["content"]
-        if 2 in sections:
-            remedies["can_appeal"] = sections[2]["content"]
-        if 3 in sections:
-            remedies["appeal_details"] = sections[3]["content"]
-        if 4 in sections:
-            remedies["first_action"] = sections[4]["content"]
-        if 5 in sections:
-            remedies["deadline"] = sections[5]["content"]
-
     return remedies
 
 
@@ -343,10 +143,10 @@ def main():
     if uploaded_file and st.button("🚀 Generate Summary"):
         with st.spinner("Processing judgment…"):
             try:
-                raw_text = extract_text_from_pdf(uploaded_file)
-                safe_text = compress_text(raw_text)
+                raw_text = core.extract_text_from_pdf(uploaded_file)
+                safe_text = core.compress_text(raw_text)
 
-                prompt = build_prompt(safe_text, language)
+                prompt = core.build_summary_prompt(safe_text, language)
 
                 # ⚡ Best multilingual model for Hindi/Bengali/Urdu
                 model_id = "meta-llama/llama-3.1-8b-instruct"
@@ -370,8 +170,8 @@ def main():
                 # -----------------------------
                 # RETRY IF ENGLISH LEAKAGE
                 # -----------------------------
-                if language.lower() != "english" and english_leakage_detected(summary):
-                    retry_prompt = build_retry_prompt(safe_text, language)
+                if language.lower() != "english" and core.english_leakage_detected(summary):
+                    retry_prompt = core.build_retry_prompt(safe_text, language)
 
                     response2 = client.chat.completions.create(
                         model=model_id,
@@ -385,7 +185,7 @@ def main():
 
                     retry_summary = response2.choices[0].message.content.strip()
 
-                    if len(retry_summary) > 0 and not english_leakage_detected(retry_summary):
+                    if len(retry_summary) > 0 and not core.english_leakage_detected(retry_summary):
                         summary = retry_summary
 
                 if not summary:
